Moving EventStoreConfiguration into Dolittle folder
