+++ Source/typescript/backend/dolittle/EventStoreConfiguration.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-import { ResourceConfiguration } from './ResourceConfiguration';
+import { ResourceConfiguration } from '../resources/ResourceConfiguration';
 
 /**
  * Represents the configuration for the Dolittle Event Store
