+++ Source/typescript/backend/resources/index.ts
@@ -1,7 +1,6 @@
 // Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-export * from './EventStoreConfiguration';
 export * from './MissingResourceConfigurationOfType';
 export * from './MissingResourceConfigurationsForTenant';
 export * from './ResourceConfiguration';
